feat(vehicle-tracking): refactor VehicleResource and VehicleTrackingController for improved trip management

+++ src/main/java/com/codeminds/edugo/vehicule/domain/model/queries/ViewRouteHistoryQuery.java
@@ -2,5 +2,5 @@ package com.codeminds.edugo.vehicule.domain.model.queries;
 
 import java.time.LocalDateTime;
 
-public record ViewRouteHistoryQuery(int vehicleId, LocalDateTime startDate, LocalDateTime endDate) {
+public record ViewRouteHistoryQuery(Long vehicleId, LocalDateTime startDate, LocalDateTime endDate) {
 }
